level builder refactoring
